Cleaning: desktop_save must not be a constant if tested
git-svn-id: file:///home/code/svn/ovd/trunk@6504 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ client/java/src/net/propero/rdp/Rdp.java
@@ -1233,7 +1233,7 @@ public class Rdp {
         order_caps[8] = 1; /* line */
         order_caps[9] = 1; /* line */
         order_caps[10] = 1; /* rect */
-        order_caps[11] = (Constants.desktop_save ? 1 : 0); /* desksave */
+        order_caps[11] = (byte) (Constants.desktop_save ? 1 : 0); /* desksave */
         order_caps[13] = 1; /* memblt */
         order_caps[14] = 1; /* triblt */
         order_caps[20] = (byte) (this.opt.polygon_ellipse_orders ? 1 : 0); /* polygon */
